+++ CHANGELOG.md
@@ -18,6 +18,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Better support for destructuring within multi assignment. (Thanks to @AlanFoster.)
 - [INTERNAL] Switch `next` test over to minitest.
 - Handle multiple arguments to `next` with a space between.
+- Handle multi-line conditional predicate (should align with keyword). (Thanks to @AndrewRayCode for the report.)
 
 ## [0.6.3] - 2019-02-18
 ### Changed